extract orderbook component to new file
