+++ src/views/DEX/DEX.tsx
@@ -4,20 +4,17 @@ import Drawer from "../../components/Drawer";
 import useBreakpoint from "hooks/useBreakpoint";
 import WidgetContainer from "views/Dashboard/components/WidgetContainer";
 import JUPAppBar from "components/JUPAppBar";
-import { Box, Button, Grid, IconButton, Link, Stack, Tab, Table, TableBody, TableCell, TableHead, TableRow, Tabs, Typography } from "@mui/material";
+import { Box, Button, Grid, IconButton, Link, Stack, Tab, Tabs, Typography } from "@mui/material";
 import JUPInput from "components/JUPInput";
-import { CSSProperties } from "@mui/styled-engine";
 import SwapVertIcon from "@mui/icons-material/SwapVert";
 import { defaultAssetList } from "utils/common/defaultAssets";
 import useAPI from "hooks/useAPI";
-import { IAsset, IGetOrdersResult } from "types/NXTAPI";
-import useBlocks from "hooks/useBlocks";
-import { NQTtoNXT } from "utils/common/NQTtoNXT";
-import { LongUnitPrecision } from "utils/common/constants";
+import { IAsset } from "types/NXTAPI";
 import OverallOrderHistory from "./components/OverallOrderHistory";
 import MyOrderHistory from "./components/MyOrderHistory";
+import OrderBook from "./components/OrderBook";
 
-const PLACEHOLDERS = {
+export const PLACEHOLDERS = {
   circulatingSupply: "123,345",
   decimals: "8",
   description:
@@ -118,105 +115,6 @@ const OrderHistory: React.FC<IOrderHistoryProps> = ({ assetId }) => {
       <TabPanel value={tabId} index={2}>
         <MyOrderHistory></MyOrderHistory>
       </TabPanel>
-    </>
-  );
-};
-
-interface IOrderbookProps {
-  assetId?: string;
-}
-
-const OrderBook: React.FC<IOrderbookProps> = ({ assetId }) => {
-  const { getOrders, getTrades } = useAPI();
-  const { blockHeight } = useBlocks();
-  const [openOrders, setOpenOrders] = useState<IGetOrdersResult>();
-
-  const bidOrderbookStyling: CSSProperties = {
-    border: "2px solid green",
-    height: "100%",
-  };
-
-  const askOrderbookStyling: CSSProperties = {
-    border: "2px solid red",
-    height: "100%",
-  };
-
-  // maps both bid and ask orders
-  const RowsMemo = useMemo(() => {
-    if (openOrders === undefined) {
-      return;
-    }
-
-    const mappedAskOrders = openOrders?.asks.map((order, index) => {
-      return (
-        <TableRow key={index}>
-          <TableCell>{NQTtoNXT(order.priceNQT).toFixed(LongUnitPrecision)}</TableCell>
-          <TableCell>{order.quantityQNT}</TableCell>
-        </TableRow>
-      );
-    });
-
-    const mappedBidOrders = openOrders?.bids.map((order, index) => {
-      return (
-        <TableRow key={index}>
-          <TableCell>{NQTtoNXT(order.priceNQT).toFixed(LongUnitPrecision)}</TableCell>
-          <TableCell>{order.quantityQNT}</TableCell>
-        </TableRow>
-      );
-    });
-
-    return { asks: mappedAskOrders, bids: mappedBidOrders };
-  }, [openOrders]);
-
-  // set the orders for the current asset
-  useEffect(() => {
-    async function fetchOrders() {
-      if (getOrders === undefined || assetId === undefined) {
-        return;
-      }
-
-      try {
-        const result = await getOrders(assetId);
-
-        if (result) {
-          setOpenOrders(result);
-        }
-      } catch (e) {
-        console.error("error while getting orders in DEX component:", e);
-        return;
-      }
-    }
-
-    fetchOrders();
-  }, [assetId, blockHeight, getOrders]);
-
-  return (
-    <>
-      <Table sx={askOrderbookStyling} size="small" padding="none">
-        <TableHead>
-          <TableCell>
-            <Typography>Price</Typography>
-          </TableCell>
-          <TableCell>
-            <Typography>Quantity</Typography>
-          </TableCell>
-        </TableHead>
-        <TableBody>{RowsMemo?.asks}</TableBody>
-      </Table>
-
-      <Typography>Last Price: {PLACEHOLDERS.lastPrice}</Typography>
-
-      <Table sx={bidOrderbookStyling} size="small" padding="none">
-        <TableHead>
-          <TableCell>
-            <Typography>Price</Typography>
-          </TableCell>
-          <TableCell>
-            <Typography>Quantity</Typography>
-          </TableCell>
-        </TableHead>
-        <TableBody>{RowsMemo?.bids}</TableBody>
-      </Table>
     </>
   );
 };
