+++ src/views/DEX/components/OrderBook/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./OrderBook";
